analysis/pyproject.toml
[tool.poetry]
name = "dht-lookup-dataset"
version = "0.1.0"
description = ""
authors = ["Dennis Trautwein <git@dtrautwein.eu>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dateutil = "^2.8.2"
durationpy = "^0.5"
numpy = "^1.22.1"
matplotlib = "^3.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
